requirements.txt
asgiref==3.8.1
cookiecutter==2.6.0
django==5.0.7
django-allauth==65.0.2
django-browser-reload==1.13.0
django-compressor==4.5.1
django-import-export==4.2.0
django-tailwind==3.8.0
ipython
openpyxl==3.1.5
pandas==2.1.4
pillow==11.0.0
psycopg[binary]==3.2.1
pulp==2.9.0
python-dotenv==0.19.2
sqlparse==0.5.1
